refactor: MainDestinationを別ファイル化
